perf(benchmarks): restore execute-only series for core scenarios
The v0.11.39 refactor renamed JrocPhasedBenchmarks to
DromaeoExecutionBenchmarks and scoped it to Scenarios/dromaeo, which
silently dropped the jroc-execute/jint-execute/okojo-execute rows for the
core scenarios (arith, array-stress, minimal, stopwatch, stopwatch-modern,
plain-object, constructed-object). Those scenarios now only publish
compile+execute totals, breaking historical execution-phase comparisons in
Supabase.

Extract the phased-execution harness into ScenarioExecutionBenchmarksBase,
reduce DromaeoExecutionBenchmarks to a scenario-directory override, and add
CoreExecutionBenchmarks (--core-execution) plus a CI matrix job so the core
scenarios publish execute-only measurements again.
