+++ .github/workflows/benchmarkdotnet-suite.yml
@@ -36,6 +36,9 @@ jobs:
           - benchmark_name: JavaScriptRuntimeBenchmarks
             mode_argument: ''
             artifact_suffix: javascript-runtime
+          - benchmark_name: CoreExecutionBenchmarks
+            mode_argument: --core-execution
+            artifact_suffix: core-execution
     
     steps:
       - name: Checkout repository
